Renamed p2p -> p2p_client
