+++ packages/fetchai/connections/p2p_client/connection.yaml
@@ -1,9 +1,9 @@
-name: p2p
+name: p2p_client
 author: fetchai
 version: 0.1.0
 license: Apache-2.0
 fingerprint: ""
-description: "The p2p connection provides a connection with the fetch.ai mail provider."
+description: "The p2p_client connection provides a connection with the fetch.ai mail provider."
 class_name: PeerToPeerConnection
 protocols: []
 restricted_to_protocols: []
