+++ tests/conftest.py
@@ -78,7 +78,7 @@ UNKNOWN_PROTOCOL_PUBLIC_ID = PublicId("unknown_author", "unknown_protocol", "0.1
 UNKNOWN_CONNECTION_PUBLIC_ID = PublicId("unknown_author", "unknown_connection", "0.1.0")
 UNKNOWN_SKILL_PUBLIC_ID = PublicId("unknown_author", "unknown_skill", "0.1.0")
 LOCAL_CONNECTION_PUBLIC_ID = PublicId("fetchai", "local", "0.1.0")
-P2P_CONNECTION_PUBLIC_ID = PublicId("fetchai", "p2p", "0.1.0")
+P2P_CONNECTION_PUBLIC_ID = PublicId("fetchai", "p2p_client", "0.1.0")
 STUB_CONNECTION_PUBLIC_ID = PublicId("fetchai", "stub", "0.1.0")
 DUMMY_PROTOCOL_PUBLIC_ID = PublicId("dummy_author", "dummy", "0.1.0")
 DUMMY_CONNECTION_PUBLIC_ID = PublicId("dummy_author", "dummy", "0.1.0")
