+++ tests/test_cli_gui/test_search.py
@@ -157,10 +157,10 @@ def test_real_search():
         == "The oef connection provides a wrapper around the OEF sdk."
     )
     i += 1
-    assert data[i]["id"] == "fetchai/p2p:0.1.0"
+    assert data[i]["id"] == "fetchai/p2p_client:0.1.0"
     assert (
         data[i]["description"]
-        == "The p2p connection provides a connection with the fetch.ai mail provider."
+        == "The p2p_client connection provides a connection with the fetch.ai mail provider."
     )
     i += 1
     assert data[i]["id"] == "fetchai/stub:0.1.0"
